resourcemanager cleanup
moved RMAN_ABORT to common.h
clean up includes
wrap extra rmanstate initialization with rmanstate_complete
create logroot during arg parsing
no need for print_cleanup_abort
move output_program_args into summary_log_setup
move find_namespace into rmanstate
